Testing Cart - saved items can be seen on cart
